docker: split tarfile code from docker-daemon code
tarfile.Source partially implements ImageSource reading from any regular
file, tarfile.Destination partially implements ImageDestination writing
to any stream.

The interactions with the Docker engine API remain in the docker-daemon:
transport.

Does not change behavior, apart from a few error messages now
generically talking about “docker tar files” instead of docker-daemon: .

[@cyphar: Rebased and re-did the splitting over several rebases.]
